+++ src/gemseo_umdo/formulations/_statistics/sampling/margin.py
@@ -42,7 +42,7 @@ class Margin(BaseSamplingEstimator):
     __standard_deviation: StandardDeviation
     """The iterative estimator of the standard deviation."""
 
-    def __init__(self, factor: float = 2.0) -> None:
+    def __init__(self, factor: float) -> None:
         """
         Args:
             factor: The factor related to the standard deviation.