+++ tests/formulations/test_pce.py
@@ -149,10 +149,6 @@ def test_variance(variance):
 def test_margin(mean, standard_deviation):
     """Check the PCE-based estimator of the margin."""
     assert_equal(
-        Margin().estimate_statistic(mean, standard_deviation),
-        mean + 2 * standard_deviation,
-    )
-    assert_equal(
         Margin(3).estimate_statistic(mean, standard_deviation),
         mean + 3 * standard_deviation,
     )